+++ global.js
@@ -9,6 +9,8 @@
  * @since 1.7
  *	- added classname substitution style.
  *	- added display_data_url option.
+ * @since 3.0
+ * 	- added debug substitution style.
  * @var object
  */
 const DEFAULT_OPTIONS = {
@@ -16,6 +18,7 @@ const DEFAULT_OPTIONS = {
 		error: 'color:red;',
 		warn: 'color:orange;',
 		info: 'color:limegreen;',
+		debug: '',
 		log: '',
 		group: 'color:mediumturquoise;border-bottom:1px dashed;cursor:pointer;',
 		number: 'background-color:dodgerblue;color:white;font-weight:bold;border-radius:0.5em;padding:0em 0.3em;',
@@ -26,29 +29,3 @@ const DEFAULT_OPTIONS = {
 	display_data_url: true
 };
 
-
-/**
- * Array.map callback
- * Removes "__proto__" and "length" properties from objects and arrays.
- * @since 1.3
- * @param mixed obj
- * @return mixed
-  */
-function cleanObjectProperties( obj ) {
-
-	// not an object ? return as-is ...
-	if ( typeof obj !== 'object' || obj === null ) return obj;
-
-	// removes length property from arrays ...
-	if ( Array.isArray(obj) ) obj = Object.assign({}, obj);
-
-	// remove annoying __proto__ property ...
-	obj.__proto__ = null;
-
-	// recurse through properties ...
-	Object.entries(obj).forEach(([ key, value ])=>{ obj[ key ] = cleanObjectProperties( value ); });
-
-	// return cloned object ...
-	return obj;
-
-}